requirements.txt
# 科技论文贡献摘要生成 - 依赖
# Python 3.10+

# 深度学习与 NLP
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# 长文档与 PDF
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# 评估指标
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.0

# 数据处理与工具
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0

# arXiv 与网络
requests>=2.31.0
arxiv>=2.0.0

# 报告/实验
wandb>=0.16.0
